Added a new methon in Network.java
The project said "Add a feature
to enable people to see a list of their friends’ friends."

I added a method in Network. It takes a User and their specific friend and finds them in the Network. If they are found, it return's the specific friend's list of friends.

Not sure if this is what he wanted. I also added a test case for it. Can you double check?